Move parseUnsubscribeHeader to index.ts
