+++ packages/inbound-email-handler/src/newsletter.ts
@@ -1,6 +1,7 @@
 import { PubSub } from '@google-cloud/pubsub'
 import { v4 as uuidv4 } from 'uuid'
 import addressparser from 'addressparser'
+import { parseUnsubscribe } from './index'
 
 const pubsub = new PubSub()
 const NEWSLETTER_EMAIL_RECEIVED_TOPIC = 'newsletterEmailReceived'
@@ -9,13 +10,6 @@ const EMAIL_FORWARDING_SENDER_ADDRESSES = [
   'Gmail Team <forwarding-noreply@google.com>',
 ]
 const CONFIRMATION_CODE_PATTERN = /^\(#\d+\)/
-const UNSUBSCRIBE_HTTP_URL_PATTERN = /<(https?:\/\/[^>]*)>/
-const UNSUBSCRIBE_MAIL_TO_PATTERN = /<mailto:([^>]*)>/
-
-interface Unsubscribe {
-  mailTo?: string
-  httpUrl?: string
-}
 
 export class NewsletterHandler {
   protected senderRegex = /NEWSLETTER_SENDER_REGEX/
@@ -48,15 +42,6 @@ export class NewsletterHandler {
     return from
   }
 
-  parseUnsubscribe(unSubHeader: string): Unsubscribe {
-    // parse list-unsubscribe header
-    // e.g. List-Unsubscribe: <https://omnivore.com/unsub>, <mailto:unsub@omnivore.com>
-    return {
-      mailTo: unSubHeader.match(UNSUBSCRIBE_MAIL_TO_PATTERN)?.[1],
-      httpUrl: unSubHeader.match(UNSUBSCRIBE_HTTP_URL_PATTERN)?.[1],
-    }
-  }
-
   async handleNewsletter(
     email: string,
     html: string,
@@ -78,7 +63,7 @@ export class NewsletterHandler {
       this.parseNewsletterUrl(postHeader, html) ||
       `${this.defaultUrl}?source=newsletters&id=${uuidv4()}`
     const author = this.parseAuthor(from)
-    const unsubscribe = this.parseUnsubscribe(unSubHeader)
+    const unsubscribe = parseUnsubscribe(unSubHeader)
     const message = {
       email,
       content: html,
